collect reductions
